业务类型修改

+++ src/main/java/com/ogc/standard/domain/CarbonBubbleOrder.java
@@ -1,5 +1,6 @@
 package com.ogc.standard.domain;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.ogc.standard.dao.base.ABaseDO;
@@ -30,7 +31,7 @@ public class CarbonBubbleOrder extends ABaseDO {
     private Date invalidDatetime;
 
     // 碳泡泡数量
-    private Integer quantity;
+    private BigDecimal quantity;
 
     // 状态（待收取、已收完、已过期）
     private String status;
@@ -108,11 +109,11 @@ public class CarbonBubbleOrder extends ABaseDO {
         this.invalidDatetime = invalidDatetime;
     }
 
-    public Integer getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
